food_genie-master/food_genie-master/requirements.txt
flask
werkzeug
python-dotenv
google-genai
requests
pillow
